Clear attributes AFTER view closes
When you close the view, keyboard state gets reset. Prior to
this I was clearing it and updating the display of the
keyboard before closing it, leading to a little bit of
flickering and redrawing right before the keyboard disappears.
After this update, it doesn't redraw the keyboard until
after it's off-screen, removing that little flicker.