Styling the settings page. And adding the launcher
Right now the launcher just takes you directly
to the settings page. I'll probably instead need
to make it take you to a "main" activity that
can tell you whether the keyboard is enabled or not.

Now that the preferences are in a fragment, perhaps
the "Main" activity could have the "active or not"
button, and also show the keyboard's preferences.