Handle bounce delay with Longs rather than messages
I forgot that Messages are not the best way to handle
a simple delay like this, because they aren't guaranteed
to happen at a specific time. For this purpose, all I really
need is the ability to keep the key from working until
a specified time.

There may be a more performant way of achieving this, though...